Initial Implementation of Scripting with Mono and C#
Currently have Log full implemented and some of Input. With an initial form of Entity that allows for testing. The script component takes a namespace and class of an C# script that inherits from Entity.

Still need to add serialization, better editor UI for implementing script component, and other classes.

Much further down the road is reflection and improvements, plus the ability to have a separate game project.